Extract activity text from TitlePrivate to separate component
Makes it reusable (will likely add it to AccountDetails too)
In rare cases we might not have presence data loaded, checks for
undefined and does not render if so.

Fixes #2104
