+++ src/title/TitlePrivate.js
@@ -4,7 +4,7 @@ import { StyleSheet, Text, View } from 'react-native';
 
 import type { User } from '../types';
 import { Avatar } from '../common';
-import { presenceToHumanTime } from '../utils/date';
+import ActivityText from './ActivityText';
 
 const styles = StyleSheet.create({
   wrapper: {
@@ -16,23 +16,17 @@ const styles = StyleSheet.create({
     fontSize: 18,
     paddingLeft: 8,
   },
-  time: {
-    fontSize: 13,
-    paddingLeft: 8,
-  },
 });
 
 type Props = {
   user: User,
   color: string,
-  presences: Object,
 };
 
 export default class TitlePrivate extends PureComponent<Props> {
   render() {
-    const { presences, user, color } = this.props;
+    const { user, color } = this.props;
     const { fullName, avatarUrl, email } = user;
-    const activity = presenceToHumanTime(presences[email]);
 
     return (
       <View style={styles.wrapper}>
@@ -41,7 +35,7 @@ export default class TitlePrivate extends PureComponent<Props> {
           <Text style={[styles.title, { color }]} numberOfLines={1} ellipsizeMode="tail">
             {fullName}
           </Text>
-          <Text style={[styles.time, { color }]}>Active {activity}</Text>
+          <ActivityText color={color} email={email} />
         </View>
       </View>
     );
