+++ src/title/TitlePrivateContainer.js
@@ -1,9 +1,8 @@
 /* @flow */
 import connectWithActions from '../connectWithActions';
-import { getPresence, getUserInPmNarrow } from '../selectors';
+import { getUserInPmNarrow } from '../selectors';
 import TitlePrivate from './TitlePrivate';
 
 export default connectWithActions((state, props) => ({
-  presences: getPresence(state),
   user: getUserInPmNarrow(props.narrow)(state),
 }))(TitlePrivate);
